requirements.txt
pandas
pymysql
datetime
sqlalchemy
flask
flask-cors
python-dotenv